Add missing GlanceAPI webhook
In general it is not allowed to create a glanceAPI instance without
defining the top-level Glance CR.
However, if for any, unknown reason a glanceAPI CR is created, it
will always fail because of the required .spec.ContainerImage field
that appears to be empty or simply not defined.
This patch introduces the glanceAPI missing webhook that covers at
the glanceAPI level the ContainerImage propagationa/definition.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>