Do not append "single" to the single glanceAPI
This is a cosmetic change that improves the user experience when the
glance Pod is deployed.
It's good to have the distinction between internal and external in
case we split the API, but if we have a single instance it doesn't
make sense to call the pod as "glance-single-api".
This patch just improves the naming when a single glanceAPI instance
is deployed.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>